Some 64-bit for ooDialog 3.2.0
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@6451 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/mark/ooDialog.beta/extensions/platform/windows/oodialog.3.2.0/oovmsg.cpp
@@ -468,8 +468,8 @@ size_t RexxEntry SendWinMsg(const char *funcname, size_t argc, CONSTRXSTRING *ar
         if ( n[2] == LB_SETCURSEL )
         {
           // at first check if it is an multiple selection lb
-          LONG style;
-          style = GetWindowLongPtr(GetDlgItem( hWnd, n[1] ), GWL_STYLE);
+          uint32_t style;
+          style = (uint32_t)GetWindowLongPtr(GetDlgItem( hWnd, n[1] ), GWL_STYLE);
 
           if ( style & LBS_MULTIPLESEL )
             if ( argv[5].strptr[0] == 'D' )
